# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for the GoChat",
        "title": "GoChat API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/me": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "description": "Modify user data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_user.ModifyUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/me/channels": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create DM channel",
                "parameters": [
                    {
                        "description": "Recipient data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_user.CreateDMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/me/channels/group": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create group DM channel",
                "parameters": [
                    {
                        "description": "Group DM data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_user.CreateDMManyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/me/guilds": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user guilds",
                "responses": {
                    "200": {
                        "description": "Guilds list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Guild"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/me/guilds/{guild_id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Leave guild",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guild id",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "406": {
                        "description": "Unable to leave your guild",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/me/guilds/{guild_id}/member": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user guild member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guild id",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guild member",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Member"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{user_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID or 'me'",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.User"
                        }
                    },
                    "400": {
                        "description": "Incorrect ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api_endpoints_user.CreateDMManyRequest": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "type": "integer"
                },
                "recipients_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "cmd_api_endpoints_user.CreateDMRequest": {
            "type": "object",
            "properties": {
                "recipient_id": {
                    "type": "integer"
                }
            }
        },
        "cmd_api_endpoints_user.ModifyUserRequest": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "avatar": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_FlameInTheDark_gochat_internal_dto.Guild": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "integer"
                },
                "public": {
                    "type": "boolean"
                }
            }
        },
        "github_com_FlameInTheDark_gochat_internal_dto.Member": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "integer"
                },
                "join_at": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_id": {
                    "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.User"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FlameInTheDark_gochat_internal_dto.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "integer"
                },
                "discriminator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "basic"
        }
    }
}